+++ src/main/java/com/pisip/jbpharma/infraestructura/repositorio/iValidacionSemaforicajpaRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pisip.jbpharma.infraestructura.persistencia.jpa.ValidacionSemaforicaEntity;
 
-public interface iValidacionSemaforicajpaRepositorio extends JpaRepository<ValidacionSemaforicaEntity, Integer> {
+public interface iValidacionSemaforicajpaRepositorio extends JpaRepository<ValidacionSemaforicaEntity, Long> {
 
 }
